implement authorization and authentication add login/register update the device logic

+++ DeviceManager/src/app/components/devices/device-list/device-list.ts
@@ -2,11 +2,12 @@ import { CommonModule } from '@angular/common';
 import { Component, computed, inject, OnInit, signal } from '@angular/core';
 import { Device as DeviceService } from '../../../shared/services/device';
 import { Toaster } from '../../../core/services/toaster';
-import { Device, DeviceType, DeviceTypeLabel } from '../../../shared/models/device.model';
+import { Device, DeviceTypeLabel } from '../../../shared/models/device.model';
 import { FormsModule } from '@angular/forms';
 import { ConfirmDialog } from '../../../shared/components/confirm-dialog/confirm-dialog';
 import { MatIcon } from '@angular/material/icon';
 import { Router } from '@angular/router';
+import { Auth } from '../../../shared/services/auth';
 
 @Component({
   selector: 'app-device-list',
@@ -16,6 +17,7 @@ import { Router } from '@angular/router';
 })
 export class DeviceList implements OnInit {
   protected readonly deviceService = inject(DeviceService);
+  protected readonly authService = inject(Auth);
   protected readonly router = inject(Router);
   private readonly toaster = inject(Toaster);
 
@@ -23,6 +25,19 @@ export class DeviceList implements OnInit {
   protected readonly deviceToDelete = signal<Device | null>(null);
 
   protected readonly DeviceType = DeviceTypeLabel;
+
+  protected readonly message = this.authService.isAdmin()
+    ? 'This device will be permanently removed from the system.'
+    : 'This device will be unassigned from you.';
+
+  readonly currentUserId = computed(() => this.authService.linkedUserId());
+  isAssignedToMe(assignedUserId?: number | null): boolean {
+    var isAssignedToMe =
+      assignedUserId !== null &&
+      assignedUserId !== undefined &&
+      assignedUserId === this.currentUserId();
+    return isAssignedToMe;
+  }
 
   // search bar logic
   protected searchTerm = signal('');
@@ -62,11 +77,20 @@ export class DeviceList implements OnInit {
   onDeleteConfirmed(): void {
     const device = this.deviceToDelete();
     if (!device) return;
-    this.deviceService.delete(device.id).subscribe({
-      next: () => {
-        this.toaster.success(`"${device.name}" deleted successfully.`);
-        this.showDeleteDialog.set(false);
-      },
-    });
+    if (this.authService.isAdmin()) {
+      this.deviceService.delete(device.id).subscribe({
+        next: () => {
+          this.toaster.success(`"${device.name}" deleted successfully.`);
+          this.showDeleteDialog.set(false);
+        },
+      });
+    } else {
+      this.deviceService.selfUnassign(device.id).subscribe({
+        next: () => {
+          this.toaster.success(`"${device.name}" returned successfully.`);
+          this.showDeleteDialog.set(false);
+        },
+      });
+    }
   }
 }
